requirements.txt
chris_plugin==0.4.0
python-chrisclient==2.12.0
pandas
pflog==1.2.26
loguru